pyproject.toml
[project]
name = "ai-portrait-generator"
version = "0.1.0"
description = "Streamlit dashboard for AI-powered portrait generation"
authors = [
    {name = "Arthur Araujo", email = "arthur@example.com"},
]
dependencies = [
    "streamlit>=1.29.0",
    "openai>=1.3.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.9"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[project.scripts]
app = "ai_portrait_generator.main:main"